cann/requirements.txt
requests>=2.31.0
docker>=7.0.0
gitpython>=3.1.40
pyyaml>=6.0
numpy>=1.24.0
mcp>=1.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
beautifulsoup4>=4.12.0
